feat: create hydra client when starting dev sever
Also get the port from ENV
