+++ package/bin/dev-server
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+/* eslint-disable node/shebang */
+require("../scripts/devServer")();
